initial draft for editors, adding perlin water step
